+++ client/src/hooks/index.ts
@@ -2,6 +2,7 @@ export * from './Messages';
 export * from './Config';
 export * from './Input';
 export * from './Conversations';
+export * from './Nav';
 
 export * from './AuthContext';
 export * from './ThemeContext';
